i no sabi the problem

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	long unsigned int counter = 0;
+	size_t counter = 0;
 
 	if (h->str == NULL)
 	{
