+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,7 +8,7 @@
 
 size_t list_len(const list_t *h)
 {
-	int counter = 0;
+	long unsigned int counter = 0;
 
 	while (h)
 	{
@@ -16,5 +16,4 @@ size_t list_len(const list_t *h)
 		counter++;
 	}
 	return (counter);
-
 }
